+++ R/dyplot.R
@@ -1,3 +1,4 @@
 dyplot <- function(x, ...) {
+  .onAttach()
   UseMethod("dyplot")
 }